Fix the new Lenz Ops mode programmer test for Z21XNetOpsModeProgrammer a different way.
